Migrated to Flutter 3.13 beta

+++ lib/services/post_service.dart
@@ -12,12 +12,16 @@ class PostService extends Service{
   resetProfilePicture() async {
     try {
       DocumentSnapshot doc = await usersRef.doc(auth.currentUser!.uid).get();
-      var user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
-      if (user.photoUrl != null) {
-        await storage.refFromURL(user.photoUrl!).delete();
-      } else {
 
-        return;
+      if(doc.exists) {
+        var user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
+
+        if (user.photoUrl != null) {
+          await storage.refFromURL(user.photoUrl!).delete();
+        } else {
+
+          return;
+        }
       }
     } catch (e) {
 
@@ -27,6 +31,7 @@ class PostService extends Service{
 
   uploadProfilePicture(File image, User user) async {
     await resetProfilePicture();
+
     String link = await uploadImage(profilePic, image);
 
     var ref = usersRef.doc(user.uid);
@@ -41,46 +46,136 @@ class PostService extends Service{
 
     for (File img in image) {
       String link = await uploadImage(posts, img);
+
       postLink.add(link);
     }
 
     DocumentSnapshot doc = await usersRef.doc(auth.currentUser!.uid).get();
 
-    UserModel user = UserModel.fromJson(
-      doc.data() as Map<String, dynamic>,
-    );
+    if(doc.exists) {
+      UserModel user = UserModel.fromJson(
+        doc.data() as Map<String, dynamic>,
+      );
 
-    var ref = postRef.doc();
+      var ref = postRef.doc();
 
-    ref.set({
-      "id": ref.id,
-      "postId": ref.id,
-      "username": user.username,
-      "ownerId": auth.currentUser!.uid,
-      "mediaUrl": postLink,
-      "description": description,
-      "location": location,
-      "timestamp": Timestamp.now(),
-      "hashtags": hashtagsList,
-      "mentions": mentionsList,
-      "type" : 'post'
-    }).catchError((e) {});
+      ref.set({
+        "id": ref.id,
+        "postId": ref.id,
+        "username": user.username,
+        "ownerId": auth.currentUser!.uid,
+        "mediaUrl": postLink,
+        "description": description,
+        "location": location,
+        "timestamp": Timestamp.now(),
+        "hashtags": hashtagsList,
+        "mentions": mentionsList,
+        "type" : 'post'
+      }).catchError((e) {});
 
-    return ref.id;
+      await addPostIdToUserCollection(ref.id);
+
+      return ref.id;
+    }
+
+    return '';
   }
 
-  addPostToHashtagsCollection(String postId, List<String> hashtagsList) async {
-    for (String hashtag in hashtagsList) {
-      await hashTagsRef.doc(hashtag).collection('posts').doc(postId).set({
-        "postId": postId,
-      });
-      await hashTagsRef.doc(hashtag).update({
-        "count": FieldValue.increment(1),
-      });
+  addPostToHashtagsCollection(String postId, List<String> hashTags) async {
+    for (String hashTag in hashTags) {
+      DocumentSnapshot doc = await hashTagsRef.doc(hashTag).get();
+
+      if (doc.exists) {
+        await hashTagsRef.doc(hashTag).update({
+          "count": FieldValue.increment(1),
+          "posts": FieldValue.arrayUnion([postId]),
+        });
+      } else {
+        await hashTagsRef.doc(hashTag).set({
+          "count": 1,
+          "posts": [postId],
+        });
+      }
+    }
+    await addOrIncrementHashtagsInUserCollection(hashTags);
+  }
+
+  addOrIncrementHashtagsInUserCollection(List<String> hashTags) async {
+    for (String hashTag in hashTags) {
+      DocumentSnapshot doc = await usersRef.doc(auth.currentUser!.uid).get();
+
+      if (doc.exists) {
+
+        await usersRef.doc(auth.currentUser!.uid).update({
+          "hashtags.$hashTag": FieldValue.increment(1),
+        });
+      }
     }
   }
 
-  uploadComment(String currentUserId, String comment, String postId,
+  decrementHashtagsInUserCollection(List<String> hashTags, int value) async {
+    for (String hashTag in hashTags) {
+      DocumentSnapshot doc = await usersRef.doc(auth.currentUser!.uid).get();
+
+      if (doc.exists) {
+        await usersRef.doc(auth.currentUser!.uid).update({
+          "hashtags.$hashTag": FieldValue.increment(-value),
+        });
+      }
+    }
+  }
+
+  addPostIdToUserCollection(String postId) async {
+    DocumentSnapshot doc = await usersRef.doc(auth.currentUser!.uid).get();
+
+    if(doc.exists) {
+      UserModel user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
+
+      if (user.postsIds!.isNotEmpty) {
+        await usersRef.doc(auth.currentUser!.uid).update({
+          "postsIds": FieldValue.arrayUnion([postId]),
+        });
+      } else {
+        await usersRef.doc(auth.currentUser!.uid).update({
+          "postsIds": [postId],
+        });
+      }
+    }
+  }
+
+  incrementUserPostCount() async {
+    var ref = usersRef.doc(auth.currentUser!.uid);
+
+    ref.update({
+      "posts": FieldValue.increment(1),
+    });
+  }
+
+  decrementUserPostCount() async {
+    var ref = usersRef.doc(auth.currentUser!.uid);
+
+    ref.update({
+      "posts": FieldValue.increment(-1),
+    });
+  }
+
+  incrementMentionsCount(String userId) async {
+    var ref = usersRef.doc(userId);
+
+    ref.update({
+      "mentionsCount": FieldValue.increment(1),
+    });
+  }
+
+  decrementMentionsCount(String userId) async {
+    var ref = usersRef.doc(userId);
+
+    ref.update({
+      "mentionsCount": FieldValue.increment(-1),
+    });
+  }
+
+  /*uploadComment(String currentUserId, String comment, String postId,
       String ownerId, String mediaUrl) async {
     DocumentSnapshot doc = await usersRef.doc(currentUserId).get();
     UserModel user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
@@ -96,6 +191,66 @@ class PostService extends Service{
       addCommentToNotification("comment", comment, user.username!, user.id!,
           postId, mediaUrl, ownerId, user.photoUrl!);
     }
+  }*/
+
+  getUserIdWithUserName(String username) async {
+    QuerySnapshot snapshot = await usersRef.where('username', isEqualTo: username).get();
+
+    if(snapshot.docs.isNotEmpty){
+
+      return snapshot.docs.first.id;
+    }
+
+    return null;
+  }
+
+  addMentionToNotification(List<String> mentions, String postId) async {
+    DocumentSnapshot doc = await usersRef.doc(auth.currentUser!.uid).get();
+
+    if(doc.exists) {
+      var user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
+
+      for (String mention in mentions) {
+        mention = mention.replaceAll("@", "");
+        String userId = await getUserIdWithUserName(mention);
+
+        DocumentSnapshot doc = await usersRef.doc(userId).get();
+
+        if (doc.exists) {
+          await notificationRef.doc(userId).collection('notifications').add({
+            "type": "mention",
+            "username": user.username,
+            "userId": user.id,
+            "profilePic": user.photoUrl,
+            "postId": postId,
+            "timestamp": Timestamp.now(),
+          });
+          await addMentionedPostIdToUserCollection(postId, userId);
+        } else {
+
+          continue;
+        }
+      }
+    }
+  }
+
+  addMentionedPostIdToUserCollection(String postId, String userId) async {
+    DocumentSnapshot doc = await usersRef.doc(userId).get();
+
+    if (doc.exists) {
+      var user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
+
+      if(user.mentionsIds!.isNotEmpty) {
+        await usersRef.doc(userId).update({
+          "mentionsIds": FieldValue.arrayUnion([postId]),
+        });
+      } else {
+        await usersRef.doc(userId).update({
+          "mentionsIds": [postId],
+        });
+      }
+    }
+    await incrementMentionsCount(userId);
   }
 
   addCommentToNotification(
@@ -104,52 +259,44 @@ class PostService extends Service{
       String username,
       String userId,
       String postId,
-      String mediaUrl,
       String ownerId,
-      String userDp) async {
-    await notificationRef.doc(ownerId).collection('notifications').add({
+      String profilePic) async {
+    await notificationRef.doc(ownerId).collection('notifications').doc(postId).set({
       "type": type,
-      "commentData": commentData,
+      "comment": commentData,
       "username": username,
       "userId": userId,
-      "userDp": userDp,
+      "profilePic": profilePic,
       "postId": postId,
-      "mediaUrl": mediaUrl,
       "timestamp": Timestamp.now(),
     });
   }
 
   addLikesToNotification(String type, String username, String userId,
-      String postId, String mediaUrl, String ownerId, String userDp) async {
-    await notificationRef
-        .doc(ownerId)
-        .collection('notifications')
-        .doc(postId)
-        .set({
+      String postId, String ownerId, String profilePic, List<dynamic> hashTags) async {
+    await notificationRef.doc(ownerId).collection('notifications').doc(postId).set({
       "type": type,
       "username": username,
       "userId": auth.currentUser!.uid,
-      "userDp": userDp,
+      "profilePic": profilePic,
       "postId": postId,
-      "mediaUrl": mediaUrl,
       "timestamp": Timestamp.now(),
     });
+    if(hashTags.isNotEmpty){
+      await addOrIncrementHashtagsInUserCollection(hashTags as List<String>);
+    }
   }
 
-  removeLikeFromNotification(
-      String ownerId, String postId, String currentUser) async {
+  removeLikeFromNotification(String ownerId, String postId, String currentUser, List<dynamic> hashTags) async {
     bool isNotMe = currentUser != ownerId;
 
     if (isNotMe) {
-      DocumentSnapshot doc = await usersRef.doc(currentUser).get();
-      UserModel user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
-      notificationRef
-          .doc(ownerId)
-          .collection('notifications')
-          .doc(postId)
-          .get()
-          .then((doc) => {
-        if (doc.exists) {doc.reference.delete()}
+      notificationRef.doc(ownerId).collection('notifications').doc(postId).get().then((doc) async {
+            if (doc.exists) {
+              doc.reference.delete();
+            }
+
+            await decrementHashtagsInUserCollection(hashTags as List<String>, 1);
       });
     }
   }
